requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
aiosqlite==0.19.0
celery==5.3.4
redis==5.0.1
huggingface-hub==0.19.4
transformers==4.35.2